get sequence renderer reversin
