+++ packages/jbrowse-web/src/plugins/DivSequenceRenderer/components/DivSequenceRendering.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 import { observer } from 'mobx-react'
 import ReactPropTypes from 'prop-types'
 
@@ -6,56 +6,57 @@ import './DivSequenceRendering.scss'
 
 import { PropTypes as CommonPropTypes } from '../../../mst-types'
 
-@observer
-class DivSequenceRendering extends Component {
-  static propTypes = {
-    region: CommonPropTypes.Region.isRequired,
-    bpPerPx: ReactPropTypes.number.isRequired,
-    features: ReactPropTypes.instanceOf(Map),
-    trackModel: ReactPropTypes.shape({
-      /** id of the currently selected feature, if any */
-      selectedFeatureId: ReactPropTypes.string,
-    }),
+function SequenceDivs({ features, region, bpPerPx, horizontallyFlipped }) {
+  let s = ''
+  for (const seq of features.values()) {
+    s += seq.seq || seq.data.seq // index.js:1452 Warning: Text content did not match. fix?
   }
 
-  static defaultProps = {
-    trackModel: {},
+  const width = (region.end - region.start) / bpPerPx
 
-    features: new Map(),
-  }
+  s = s.split('')
+  if (horizontallyFlipped) s = s.reverse()
 
-  render() {
-    const { region, bpPerPx, features } = this.props
-
-    let s = ''
-    for (const seq of features.values()) {
-      s += seq.seq || seq.data.seq // index.js:1452 Warning: Text content did not match. fix?
-    }
-
-    const width = (region.end - region.start) / bpPerPx
-
-    return (
-      <div className="DivSequenceRendering">
-        {bpPerPx >= 1 ? (
-          <div className="blur">Zoom in to see sequence</div>
-        ) : (
-          <div>
-            {s.split('').map((letter, iter) => (
-              <div
-                /* eslint-disable-next-line */
-                key={`${region.start}-${iter}`}
-                style={{
-                  width: `${width / s.length}px`,
-                }}
-                className={`base base-${letter.toLowerCase()}`}
-              >
-                {bpPerPx < 0.1 ? letter : '\u00A0'}
-              </div>
-            ))}
-          </div>
-        )}
-      </div>
-    )
-  }
+  return (
+    <>
+      {s.map((letter, iter) => (
+        <div
+          /* eslint-disable-next-line */
+          key={`${region.start}-${iter}`}
+          style={{
+            width: `${width / s.length}px`,
+          }}
+          className={`base base-${letter.toLowerCase()}`}
+        >
+          {bpPerPx < 0.1 ? letter : '\u00A0'}
+        </div>
+      ))}
+    </>
+  )
 }
-export default DivSequenceRendering
+
+SequenceDivs.propTypes = {
+  region: CommonPropTypes.Region.isRequired,
+  bpPerPx: ReactPropTypes.number.isRequired,
+  features: ReactPropTypes.instanceOf(Map),
+  horizontallyFlipped: ReactPropTypes.bool,
+}
+SequenceDivs.defaultProps = {
+  features: new Map(),
+  horizontallyFlipped: false,
+}
+
+function DivSequenceRendering(props) {
+  const { bpPerPx } = props
+  return (
+    <div className="DivSequenceRendering">
+      {bpPerPx >= 1 ? (
+        <div className="blur">Zoom in to see sequence</div>
+      ) : (
+        <SequenceDivs {...props} />
+      )}
+    </div>
+  )
+}
+
+export default observer(DivSequenceRendering)
