+++ packages/jbrowse-web/src/plugins/SequenceTrack/configSchema.js
@@ -1,24 +1,15 @@
 import { ConfigurationSchema } from '../../configuration'
 import { BaseTrackConfig as LinearGenomeTrackConfig } from '../LinearGenomeView/models/baseTrack'
 
-export default pluginManager => {
-  const SequenceRendererConfigSchema = pluginManager.getRendererType(
-    'DivSequenceRenderer',
-  ).configSchema
+export default pluginManager =>
+  // const SequenceRendererConfigSchema = pluginManager.getRendererType(
+  //   'DivSequenceRenderer',
+  // ).configSchema
 
-  return ConfigurationSchema(
+  ConfigurationSchema(
     'SequenceTrack',
     {
       adapter: pluginManager.pluggableConfigSchemaType('adapter'),
-      defaultRendering: {
-        type: 'string',
-        defaultValue: 'div',
-      },
-
-      renderers: ConfigurationSchema('RenderersConfiguration', {
-        PileupRenderer: SequenceRendererConfigSchema,
-      }),
     },
     { baseConfiguration: LinearGenomeTrackConfig, explicitlyTyped: true },
   )
-}
